+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
@@ -14,7 +14,7 @@ import {
   JSONParser as healthChecks
 } from "./serviceForm/JSONReducers/HealthChecks";
 import { JSONParser as labels } from "./serviceForm/JSONReducers/Labels";
-import { JSONParser as volumes } from "./serviceForm/JSONReducers/LocalVolumes";
+import { JSONParser as volumes } from "./serviceForm/JSONReducers/Volumes";
 import {
   JSONParser as unknownVolumes
 } from "./serviceForm/JSONReducers/UnknownVolumes";
