+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Volumes.js
@@ -1,12 +1,13 @@
 import { parseIntValue } from "#SRC/js/utils/ReducerUtil";
 import { omit } from "#SRC/js/utils/Util";
 import { ADD_ITEM, REMOVE_ITEM, SET } from "#SRC/js/constants/TransactionTypes";
+import Transaction from "#SRC/js/structs/Transaction";
 
 import ContainerConstants from "../../../constants/ContainerConstants";
 
 const { type: { MESOS, DOCKER } } = ContainerConstants;
 
-const mapLocalVolumes = function(volume) {
+const mapVolumes = function(volume) {
   if (volume.type === "PERSISTENT") {
     return {
       persistent: volume.persistent,
@@ -113,7 +114,7 @@ function reduceVolumes(state, { type, path, value }) {
        * mapping them to the common structure
        */
       return [].concat(
-        this.volumes.map(mapLocalVolumes),
+        this.volumes.map(mapVolumes),
         this.externalVolumes,
         this.unknownVolumes
       );
@@ -164,7 +165,7 @@ function reduceVolumes(state, { type, path, value }) {
       }
 
       return [].concat(
-        this.volumes.map(mapLocalVolumes),
+        this.volumes.map(mapVolumes),
         this.externalVolumes,
         this.unknownVolumes
       );
@@ -192,12 +193,80 @@ function reduceVolumes(state, { type, path, value }) {
   }
 
   return [].concat(
-    this.volumes.map(mapLocalVolumes),
+    this.volumes.map(mapVolumes),
     this.externalVolumes.map(mapExternalVolumes.bind(this)),
     this.unknownVolumes
   );
 }
 
 module.exports = {
-  JSONReducer: reduceVolumes
+  JSONReducer: reduceVolumes,
+  JSONParser(state) {
+    if (state.container == null || state.container.volumes == null) {
+      return [];
+    }
+
+    return state.container.volumes
+      .filter(item => item.hostPath != null || item.persistent != null)
+      .reduce(function(memo, item, index) {
+        /**
+         * For the volumes we have a special case as all the volumes
+         * are present in the `container.volumes` But in this parser we only
+         * want to parse the local volumes. which means that we first filter
+         * those and only keep local volumes (decision based on if
+         * persistent or hostPath is set). After that we do get all the values even
+         * stuff which we do not handle in the form yet. These steps are:
+         * 1) Add a new Item to the path with the index equal to index.
+         * 2) Set the size from `volume.persistent.size`on the path
+         *    `volumes.${index}.size`.
+         * 3) Set the containerPath from `volume.containerPath on the path
+         *    `volumes.${index}.containerPath`
+         * 4) Set the mode from `volume.mode` on the path
+         *    `volumes.${index}.mode`
+         */
+        memo.push(new Transaction(["volumes"], item, ADD_ITEM));
+
+        if (item.persistent != null && item.persistent.size != null) {
+          memo.push(
+            new Transaction(["volumes", index, "type"], "PERSISTENT", SET)
+          );
+
+          memo.push(
+            new Transaction(
+              ["volumes", index, "size"],
+              item.persistent.size,
+              SET
+            )
+          );
+        } else {
+          memo.push(new Transaction(["volumes", index, "type"], "HOST", SET));
+
+          memo.push(
+            new Transaction(
+              ["volumes", index, "hostPath"],
+              item.hostPath || "",
+              SET
+            )
+          );
+        }
+
+        if (item.containerPath != null) {
+          memo.push(
+            new Transaction(
+              ["volumes", index, "containerPath"],
+              item.containerPath,
+              SET
+            )
+          );
+        }
+
+        if (item.mode != null) {
+          memo.push(
+            new Transaction(["volumes", index, "mode"], item.mode, SET)
+          );
+        }
+
+        return memo;
+      }, []);
+  }
 };
