+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/Volumes-test.js
@@ -416,4 +416,142 @@ describe("Volumes", function() {
       ]);
     });
   });
+  describe("Volumes", function() {
+    describe("#JSONParser", function() {
+      it("should return an empty array", function() {
+        expect(Volumes.JSONParser({})).toEqual([]);
+      });
+
+      it("should return an empty array if only external volumes are present", function() {
+        const state = {
+          container: {
+            volumes: [
+              {
+                containerPath: "/mnt/volume",
+                external: {
+                  name: "someVolume",
+                  provider: "dvdi",
+                  options: {
+                    "dvdi/driver": "rexray"
+                  }
+                },
+                mode: "RW"
+              }
+            ]
+          }
+        };
+        expect(Volumes.JSONParser(state)).toEqual([]);
+      });
+
+      it("should contain the transaction for one local volume", function() {
+        const state = {
+          container: {
+            volumes: [
+              {
+                containerPath: "/dev/null",
+                persistent: { size: 1024 },
+                mode: "RW"
+              }
+            ]
+          }
+        };
+        expect(Volumes.JSONParser(state)).toEqual([
+          {
+            type: ADD_ITEM,
+            value: {
+              containerPath: "/dev/null",
+              persistent: { size: 1024 },
+              mode: "RW"
+            },
+            path: ["volumes"]
+          },
+          { type: SET, value: "PERSISTENT", path: ["volumes", 0, "type"] },
+          { type: SET, value: 1024, path: ["volumes", 0, "size"] },
+          {
+            type: SET,
+            value: "/dev/null",
+            path: ["volumes", 0, "containerPath"]
+          },
+          { type: SET, value: "RW", path: ["volumes", 0, "mode"] }
+        ]);
+      });
+
+      it("should exclude the external volumes", function() {
+        const state = {
+          container: {
+            volumes: [
+              {
+                containerPath: "/mnt/volume",
+                external: {
+                  name: "someVolume",
+                  provider: "dvdi",
+                  options: {
+                    "dvdi/driver": "rexray"
+                  }
+                },
+                mode: "RW"
+              },
+              {
+                containerPath: "/dev/null",
+                persistent: { size: 1024 },
+                mode: "RW"
+              }
+            ]
+          }
+        };
+        expect(Volumes.JSONParser(state)).toEqual([
+          {
+            type: ADD_ITEM,
+            value: {
+              containerPath: "/dev/null",
+              persistent: { size: 1024 },
+              mode: "RW"
+            },
+            path: ["volumes"]
+          },
+          { type: SET, value: "PERSISTENT", path: ["volumes", 0, "type"] },
+          { type: SET, value: 1024, path: ["volumes", 0, "size"] },
+          {
+            type: SET,
+            value: "/dev/null",
+            path: ["volumes", 0, "containerPath"]
+          },
+          { type: SET, value: "RW", path: ["volumes", 0, "mode"] }
+        ]);
+      });
+
+      it("should include a unknown value for modes", function() {
+        const state = {
+          container: {
+            volumes: [
+              {
+                containerPath: "/dev/null",
+                persistent: { size: 1024 },
+                mode: "READ"
+              }
+            ]
+          }
+        };
+        expect(Volumes.JSONParser(state)).toEqual([
+          {
+            type: ADD_ITEM,
+            value: {
+              containerPath: "/dev/null",
+              persistent: { size: 1024 },
+              mode: "READ"
+            },
+            path: ["volumes"]
+          },
+          { type: SET, value: "PERSISTENT", path: ["volumes", 0, "type"] },
+          { type: SET, value: 1024, path: ["volumes", 0, "size"] },
+          {
+            type: SET,
+            value: "/dev/null",
+            path: ["volumes", 0, "containerPath"]
+          },
+          { type: SET, value: "READ", path: ["volumes", 0, "mode"] }
+        ]);
+      });
+    });
+  });
 });
